+++ src/Ar/ATN/Director.cpp
@@ -2,9 +2,9 @@
  * File: Director.cpp
  * Copyright (c) 2023 Loupe
  * https://loupe.team
- * 
+ *
  * This file is part of All Together Now - ATN, licensed under the MIT License.
- * 
+ *
  */
 
 #include <iostream>
@@ -68,24 +68,6 @@ Director::~Director()
 
 }
 
-unsigned int Director::countActiveThreads(){
-    return threads.size();
-}
-
-void Director::addBehavior( const std::string action,  AtnAPI_typ* api, void *_pParameters, size_t _sParameters, const std::string& taskName ){
-
-    auto it = actions.find(action);
-
-    if (it != actions.end()){
-        it->second.subscribe( api,_pParameters, _sParameters, taskName);
-    }
-    else{
-        Action newAction(action);
-        newAction.subscribe(api,_pParameters, _sParameters, taskName);
-        actions.insert( std::pair<std::string, Action>(action, newAction));
-    }
-}
-
 void Director::addState( const std::string state,  AtnAPIState_typ* api, void *_pParameters, size_t _sParameters, const std::string& taskName ){
 
     auto it = states.find(state);
@@ -144,7 +126,7 @@ void Director::addResourceBool( const std::string state, const std::string name,
 		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
-		
+
 
 
 void Director::addStateBool( const std::string state, const std::string name, bool* value,  void *_pParameters, size_t _sParameters, const std::string& taskName){
@@ -203,23 +185,6 @@ void Director::addCommandPLCOpen( const std::string command, const std::string m
 	}
 }
 
-void Director::executeAction( const std::string action,  AtnApiStatus_typ* _pStatus, void *_pParameters, size_t _sParameters){
-
-    auto it = actions.find(action);
-
-    if (it != actions.end()){
-        threads.push_back(it->second); 
-        ///TODO: check if it's OK to start
-        threads.back().start( _pStatus, _pParameters, _sParameters );
-    }
-    else{
-		if( this->outstream ){
-			*this->outstream << "Action not found\n";
-		}
-    }
-}
-
-
 bool Director::executeCommand( const std::string command ){
 
     auto it = commands.find(command);
@@ -230,7 +195,7 @@ bool Director::executeCommand( const std::string command ){
     }
     else{
 		if( this->outstream ){
-			*this->outstream << "Action not found\n";
+			*this->outstream << "Command not found\n";
 		}
 		return false;
 	}
@@ -245,7 +210,7 @@ void Director::resetCommand( const std::string command ){
     }
     else{
 		if( this->outstream ){
-			*this->outstream << "Action not found\n";
+			*this->outstream << "Command not found\n";
 		}
     }
 }
@@ -254,7 +219,7 @@ State * Director::getState( const std::string state ){
     auto it = states.find(state);
 
     if (it != states.end()){
-        return &it->second;        
+        return &it->second;
     }
     else{
         return 0;
@@ -265,7 +230,7 @@ State * Director::getCommand( const std::string cmd ){
     auto it = commands.find(cmd);
 
     if (it != commands.end()){
-        return &it->second;        
+        return &it->second;
     }
     else{
         return 0;
@@ -326,11 +291,6 @@ unsigned int Director::removeRegistration( const std::string& name, const std::s
 		removed += c->second.removeTask(taskName);
 	}
 
-	auto a = actions.find(name);
-	if( a != actions.end() ){
-		removed += a->second.removeTask(taskName);
-	}
-
 	return removed;
 }
 
@@ -347,49 +307,10 @@ unsigned int Director::removeAllForTask( const std::string& taskName ){
 	for( auto &kv : commands ){
 		removed += kv.second.removeTask(taskName);
 	}
-	for( auto &kv : actions ){
-		removed += kv.second.removeTask(taskName);
-	}
-	//In-flight actions hold copies of the registered behaviors.
-	// Sweep them too, but do not count them as additional registrations.
-	for( auto &thread : threads ){
-		thread.removeTask(taskName);
-	}
 
 	return removed;
 }
 
-void Director::cyclic(){
-
-    for (auto thread = threads.begin(); thread != threads.end(); ){
-        //Run thread
-        if( thread->update() ){
-            if( this->outstream ){
-                thread->print( *this->outstream );
-            }
-            //If it's done, remove it and advance to the next thread. erase()
-            //returns the iterator following the removed element; reusing the
-            //old iterator after erase() is undefined behavior.
-            thread = threads.erase( thread );
-        }
-        else {
-            ++thread;
-        }
-    }
-}
-
-void Director::printState( std::ostream &out ){
-    for (auto thread = threads.begin(); thread != threads.end(); ++thread){
-        thread->print( out );   
-    }
-}
-
-void Director::printActions( std::ostream &out ){
-    out << "\nActions:" << "\n";
-    for( auto action : actions ){
-        out << action.first << "\n";
-    }
-}
 void Director::printStates( std::ostream &out ){
     out << "\nStates:" << "\n";
     for( auto state : states ){
@@ -419,7 +340,7 @@ void Director::printSystemJson( std::ostream &out ){
 	out << "\"commands\":[";
 	comma = 0;
 	for( auto command : commands ){
-		if(comma) out << ","; 
+		if(comma) out << ",";
 		out << "\"" <<command.first << "\"";
 		comma = 1;
 	}
@@ -440,5 +361,5 @@ void Director::printSystemJson( std::ostream &out ){
 		comma = 1;
 	}
 	out << "]";
-	out << "}";	
+	out << "}";
 }